backend/langchain-python/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
aiohttp==3.9.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
langchain==0.3.25
langchain-openai==0.2.14
langgraph>=0.3.0,<0.4.0
langgraph-prebuilt==0.1.8
